We would like to refactor the CatalystQueryPlanVisitor and separate it into two distinct visitors:
Plan Visitor ( which extends AbstractNodeVisitor<LogicalPlan, CatalystPlanContext> )
Expression Visitor (which extends AbstractNodeVisitor<Expression, CatalystPlanContext>)
This would match the existing PPL AST visitors composition:

AstBuilder ( which extends OpenSearchPPLParserBaseVisitor)
AstExpressionBuilder ( which extends OpenSearchPPLParserBaseVisitor )
In addition unify the ppl utils classes to match one of the following naming:

*Transformer - transforms PPL (logical) expressions into Spark (logical) expressions
*Utils - utility class
