+++ ppl-spark-integration/src/main/java/org/opensearch/sql/ppl/CatalystQueryPlanVisitor.java
@@ -90,11 +90,11 @@ import org.opensearch.sql.ast.tree.UnresolvedPlan;
 import org.opensearch.sql.ast.tree.Window;
 import org.opensearch.sql.common.antlr.SyntaxCheckException;
 import org.opensearch.sql.expression.function.SerializableUdf;
-import org.opensearch.sql.ppl.utils.AggregatorTranslator;
-import org.opensearch.sql.ppl.utils.BuiltinFunctionTranslator;
+import org.opensearch.sql.ppl.utils.AggregatorTransformer;
+import org.opensearch.sql.ppl.utils.BuiltinFunctionTransformer;
 import org.opensearch.sql.ppl.utils.ComparatorTransformer;
 import org.opensearch.sql.ppl.utils.FieldSummaryTransformer;
-import org.opensearch.sql.ppl.utils.ParseStrategy;
+import org.opensearch.sql.ppl.utils.ParseTransformer;
 import org.opensearch.sql.ppl.utils.SortUtils;
 import org.opensearch.sql.ppl.utils.WindowSpecTransformer;
 import scala.Option;
@@ -136,20 +136,16 @@ import static scala.collection.JavaConverters.seqAsJavaList;
  */
 public class CatalystQueryPlanVisitor extends AbstractNodeVisitor<LogicalPlan, CatalystPlanContext> {
 
-    private final ExpressionAnalyzer expressionAnalyzer;
+    private final CatalystExpressionVisitor expressionAnalyzer;
 
     public CatalystQueryPlanVisitor() {
-        this.expressionAnalyzer = new ExpressionAnalyzer();
+        this.expressionAnalyzer = new CatalystExpressionVisitor(this);
     }
 
     public LogicalPlan visit(Statement plan, CatalystPlanContext context) {
         return plan.accept(this, context);
     }
-
-    public LogicalPlan visitSubSearch(UnresolvedPlan plan, CatalystPlanContext context) {
-        return plan.accept(this, context);
-    }
-
+    
     /**
      * Handle Query Statement.
      */
@@ -478,7 +474,7 @@ public class CatalystQueryPlanVisitor extends AbstractNodeVisitor<LogicalPlan, C
         ParseMethod parseMethod = node.getParseMethod();
         java.util.Map<String, Literal> arguments = node.getArguments();
         String pattern = (String) node.getPattern().getValue();
-        return ParseStrategy.visitParseCommand(node, sourceField, parseMethod, arguments, pattern, context);
+        return ParseTransformer.visitParseCommand(node, sourceField, parseMethod, arguments, pattern, context);
     }
 
     @Override
@@ -570,339 +566,6 @@ public class CatalystQueryPlanVisitor extends AbstractNodeVisitor<LogicalPlan, C
             } else {
                 return retainMultipleDuplicateEvents(node, allowedDuplication, expressionAnalyzer, context);
             }
-        }
-    }
-
-    /**
-     * Expression Analyzer.
-     */
-    public class ExpressionAnalyzer extends AbstractNodeVisitor<Expression, CatalystPlanContext> {
-
-        public Expression analyze(UnresolvedExpression unresolved, CatalystPlanContext context) {
-            return unresolved.accept(this, context);
-        }
-
-        @Override
-        public Expression visitLiteral(Literal node, CatalystPlanContext context) {
-            return context.getNamedParseExpressions().push(new org.apache.spark.sql.catalyst.expressions.Literal(
-                    translate(node.getValue(), node.getType()), translate(node.getType())));
-        }
-
-        /**
-         * generic binary (And, Or, Xor , ...) arithmetic expression resolver
-         *
-         * @param node
-         * @param transformer
-         * @param context
-         * @return
-         */
-        public Expression visitBinaryArithmetic(BinaryExpression node, BiFunction<Expression, Expression, Expression> transformer, CatalystPlanContext context) {
-            node.getLeft().accept(this, context);
-            Optional<Expression> left = context.popNamedParseExpressions();
-            node.getRight().accept(this, context);
-            Optional<Expression> right = context.popNamedParseExpressions();
-            if (left.isPresent() && right.isPresent()) {
-                return transformer.apply(left.get(), right.get());
-            } else if (left.isPresent()) {
-                return context.getNamedParseExpressions().push(left.get());
-            } else if (right.isPresent()) {
-                return context.getNamedParseExpressions().push(right.get());
-            }
-            return null;
-
-        }
-
-        @Override
-        public Expression visitAnd(And node, CatalystPlanContext context) {
-            return visitBinaryArithmetic(node,
-                    (left, right) -> context.getNamedParseExpressions().push(new org.apache.spark.sql.catalyst.expressions.And(left, right)), context);
-        }
-
-        @Override
-        public Expression visitOr(Or node, CatalystPlanContext context) {
-            return visitBinaryArithmetic(node,
-                    (left, right) -> context.getNamedParseExpressions().push(new org.apache.spark.sql.catalyst.expressions.Or(left, right)), context);
-        }
-
-        @Override
-        public Expression visitXor(Xor node, CatalystPlanContext context) {
-            return visitBinaryArithmetic(node,
-                    (left, right) -> context.getNamedParseExpressions().push(new org.apache.spark.sql.catalyst.expressions.BitwiseXor(left, right)), context);
-        }
-
-        @Override
-        public Expression visitNot(Not node, CatalystPlanContext context) {
-            node.getExpression().accept(this, context);
-            Optional<Expression> arg = context.popNamedParseExpressions();
-            return arg.map(expression -> context.getNamedParseExpressions().push(new org.apache.spark.sql.catalyst.expressions.Not(expression))).orElse(null);
-        }
-
-        @Override
-        public Expression visitSpan(Span node, CatalystPlanContext context) {
-            node.getField().accept(this, context);
-            Expression field = (Expression) context.popNamedParseExpressions().get();
-            node.getValue().accept(this, context);
-            Expression value = (Expression) context.popNamedParseExpressions().get();
-            return context.getNamedParseExpressions().push(window(field, value, node.getUnit()));
-        }
-
-        @Override
-        public Expression visitAggregateFunction(AggregateFunction node, CatalystPlanContext context) {
-            node.getField().accept(this, context);
-            Expression arg = (Expression) context.popNamedParseExpressions().get();
-            Expression aggregator = AggregatorTranslator.aggregator(node, arg);
-            return context.getNamedParseExpressions().push(aggregator);
-        }
-
-        @Override
-        public Expression visitCompare(Compare node, CatalystPlanContext context) {
-            analyze(node.getLeft(), context);
-            Optional<Expression> left = context.popNamedParseExpressions();
-            analyze(node.getRight(), context);
-            Optional<Expression> right = context.popNamedParseExpressions();
-            if (left.isPresent() && right.isPresent()) {
-                Predicate comparator = ComparatorTransformer.comparator(node, left.get(), right.get());
-                return context.getNamedParseExpressions().push((org.apache.spark.sql.catalyst.expressions.Expression) comparator);
-            }
-            return null;
-        }
-
-        @Override
-        public Expression visitQualifiedName(QualifiedName node, CatalystPlanContext context) {
-            List<UnresolvedRelation> relation = findRelation(context.traversalContext());
-            if (!relation.isEmpty()) {
-                Optional<QualifiedName> resolveField = resolveField(relation, node, context.getRelations());
-                return resolveField.map(qualifiedName -> context.getNamedParseExpressions().push(UnresolvedAttribute$.MODULE$.apply(seq(qualifiedName.getParts()))))
-                    .orElse(resolveQualifiedNameWithSubqueryAlias(node, context));
-            }
-            return context.getNamedParseExpressions().push(UnresolvedAttribute$.MODULE$.apply(seq(node.getParts())));
-        }
-
-        /**
-         * Resolve the qualified name with subquery alias: <br/>
-         * - subqueryAlias1.joinKey = subqueryAlias2.joinKey <br/>
-         * - tableName1.joinKey = subqueryAlias2.joinKey <br/>
-         * - subqueryAlias1.joinKey = tableName2.joinKey <br/>
-         */
-        private Expression resolveQualifiedNameWithSubqueryAlias(QualifiedName node, CatalystPlanContext context) {
-            if (node.getPrefix().isPresent() &&
-                context.traversalContext().peek() instanceof org.apache.spark.sql.catalyst.plans.logical.SubqueryAlias) {
-                if (context.getSubqueryAlias().stream().map(p -> (org.apache.spark.sql.catalyst.plans.logical.SubqueryAlias) p)
-                    .anyMatch(a -> a.alias().equalsIgnoreCase(node.getPrefix().get().toString()))) {
-                    return context.getNamedParseExpressions().push(UnresolvedAttribute$.MODULE$.apply(seq(node.getParts())));
-                } else if (context.getRelations().stream().map(p -> (UnresolvedRelation) p)
-                    .anyMatch(a -> a.tableName().equalsIgnoreCase(node.getPrefix().get().toString()))) {
-                    return context.getNamedParseExpressions().push(UnresolvedAttribute$.MODULE$.apply(seq(node.getParts())));
-                }
-            }
-            return null;
-        }
-
-        @Override
-        public Expression visitCorrelationMapping(FieldsMapping node, CatalystPlanContext context) {
-            return node.getChild().stream().map(expression ->
-                    visitCompare((Compare) expression, context)
-            ).reduce(org.apache.spark.sql.catalyst.expressions.And::new).orElse(null);
-        }
-
-        @Override
-        public Expression visitAllFields(AllFields node, CatalystPlanContext context) {
-            context.getNamedParseExpressions().push(UnresolvedStar$.MODULE$.apply(Option.<Seq<String>>empty()));
-            return context.getNamedParseExpressions().peek();
-        }
-
-        @Override
-        public Expression visitAlias(Alias node, CatalystPlanContext context) {
-            node.getDelegated().accept(this, context);
-            Expression arg = context.popNamedParseExpressions().get();
-            return context.getNamedParseExpressions().push(
-                    org.apache.spark.sql.catalyst.expressions.Alias$.MODULE$.apply(arg,
-                            node.getAlias() != null ? node.getAlias() : node.getName(),
-                            NamedExpression.newExprId(),
-                            seq(new java.util.ArrayList<String>()),
-                            Option.empty(),
-                            seq(new java.util.ArrayList<String>())));
-        }
-
-        @Override
-        public Expression visitEval(Eval node, CatalystPlanContext context) {
-            throw new IllegalStateException("Not Supported operation : Eval");
-        }
-
-        @Override
-        public Expression visitFunction(Function node, CatalystPlanContext context) {
-            List<Expression> arguments =
-                    node.getFuncArgs().stream()
-                            .map(
-                                    unresolvedExpression -> {
-                                        var ret = analyze(unresolvedExpression, context);
-                                        if (ret == null) {
-                                            throw new UnsupportedOperationException(
-                                                    String.format("Invalid use of expression %s", unresolvedExpression));
-                                        } else {
-                                            return context.popNamedParseExpressions().get();
-                                        }
-                                    })
-                            .collect(Collectors.toList());
-            Expression function = BuiltinFunctionTranslator.builtinFunction(node, arguments);
-            return context.getNamedParseExpressions().push(function);
-        }
-
-        @Override
-        public Expression visitIsEmpty(IsEmpty node, CatalystPlanContext context) {
-            Stack<Expression> namedParseExpressions = new Stack<>();
-            namedParseExpressions.addAll(context.getNamedParseExpressions());
-            Expression expression = visitCase(node.getCaseValue(), context);
-            namedParseExpressions.add(expression);
-            context.setNamedParseExpressions(namedParseExpressions);
-            return expression;
-        }
-
-        @Override
-        public Expression visitFillNull(FillNull fillNull, CatalystPlanContext context) {
-            throw new IllegalStateException("Not Supported operation : FillNull");
-        }
-
-        @Override
-        public Expression visitInterval(Interval node, CatalystPlanContext context) {
-            throw new IllegalStateException("Not Supported operation : Interval");
-        }
-
-        @Override
-        public Expression visitDedupe(Dedupe node, CatalystPlanContext context) {
-            throw new IllegalStateException("Not Supported operation : Dedupe");
-        }
-
-        @Override
-        public Expression visitIn(In node, CatalystPlanContext context) {
-            node.getField().accept(this, context);
-            Expression value = context.popNamedParseExpressions().get();
-            List<Expression> list = node.getValueList().stream().map( expression -> {
-                expression.accept(this, context);
-                return context.popNamedParseExpressions().get();
-            }).collect(Collectors.toList());
-            return context.getNamedParseExpressions().push(In$.MODULE$.apply(value, seq(list)));
-        }
-
-        @Override
-        public Expression visitKmeans(Kmeans node, CatalystPlanContext context) {
-            throw new IllegalStateException("Not Supported operation : Kmeans");
-        }
-
-        @Override
-        public Expression visitCase(Case node, CatalystPlanContext context) {
-            Stack<Expression> initialNameExpressions = new Stack<>();
-            initialNameExpressions.addAll(context.getNamedParseExpressions());
-            analyze(node.getElseClause(), context);
-            Expression elseValue = context.getNamedParseExpressions().pop();
-            List<Tuple2<Expression, Expression>> whens = new ArrayList<>();
-            for (When when : node.getWhenClauses()) {
-                if (node.getCaseValue() == null) {
-                    whens.add(
-                            new Tuple2<>(
-                                    analyze(when.getCondition(), context),
-                                    analyze(when.getResult(), context)
-                            )
-                    );
-                } else {
-                    // Merge case value and condition (compare value) into a single equal condition
-                    Compare compare = new Compare(EQUAL.getName().getFunctionName(), node.getCaseValue(), when.getCondition());
-                    whens.add(
-                            new Tuple2<>(
-                                    analyze(compare, context), analyze(when.getResult(), context)
-                            )
-                    );
-                }
-                context.retainAllNamedParseExpressions(e -> e);
-            }
-            context.setNamedParseExpressions(initialNameExpressions);
-            return context.getNamedParseExpressions().push(new CaseWhen(seq(whens), Option.apply(elseValue)));
-        }
-
-        @Override
-        public Expression visitRareTopN(RareTopN node, CatalystPlanContext context) {
-            throw new IllegalStateException("Not Supported operation : RareTopN");
-        }
-
-        @Override
-        public Expression visitWindowFunction(WindowFunction node, CatalystPlanContext context) {
-            throw new IllegalStateException("Not Supported operation : WindowFunction");
-        }
-
-        @Override
-        public Expression visitInSubquery(InSubquery node, CatalystPlanContext outerContext) {
-            CatalystPlanContext innerContext = new CatalystPlanContext();
-            visitExpressionList(node.getChild(), innerContext);
-            Seq<Expression> values = innerContext.retainAllNamedParseExpressions(p -> p);
-            UnresolvedPlan outerPlan = node.getQuery();
-            LogicalPlan subSearch = CatalystQueryPlanVisitor.this.visitSubSearch(outerPlan, innerContext);
-            Expression inSubQuery = InSubquery$.MODULE$.apply(
-                values,
-                ListQuery$.MODULE$.apply(
-                    subSearch,
-                    seq(new java.util.ArrayList<Expression>()),
-                    NamedExpression.newExprId(),
-                    -1,
-                    seq(new java.util.ArrayList<Expression>()),
-                    Option.empty()));
-            return outerContext.getNamedParseExpressions().push(inSubQuery);
-        }
-
-        @Override
-        public Expression visitScalarSubquery(ScalarSubquery node, CatalystPlanContext context) {
-            CatalystPlanContext innerContext = new CatalystPlanContext();
-            UnresolvedPlan outerPlan = node.getQuery();
-            LogicalPlan subSearch = CatalystQueryPlanVisitor.this.visitSubSearch(outerPlan, innerContext);
-            Expression scalarSubQuery = ScalarSubquery$.MODULE$.apply(
-                subSearch,
-                seq(new java.util.ArrayList<Expression>()),
-                NamedExpression.newExprId(),
-                seq(new java.util.ArrayList<Expression>()),
-                Option.empty(),
-                Option.empty());
-            return context.getNamedParseExpressions().push(scalarSubQuery);
-        }
-
-        @Override
-        public Expression visitExistsSubquery(ExistsSubquery node, CatalystPlanContext context) {
-            CatalystPlanContext innerContext = new CatalystPlanContext();
-            UnresolvedPlan outerPlan = node.getQuery();
-            LogicalPlan subSearch = CatalystQueryPlanVisitor.this.visitSubSearch(outerPlan, innerContext);
-            Expression existsSubQuery = Exists$.MODULE$.apply(
-                subSearch,
-                seq(new java.util.ArrayList<Expression>()),
-                NamedExpression.newExprId(),
-                seq(new java.util.ArrayList<Expression>()),
-                Option.empty());
-            return context.getNamedParseExpressions().push(existsSubQuery);
-        }
-
-        @Override
-        public Expression visitBetween(Between node, CatalystPlanContext context) {
-            Expression value = analyze(node.getValue(), context);
-            Expression lower = analyze(node.getLowerBound(), context);
-            Expression upper = analyze(node.getUpperBound(), context);
-            context.retainAllNamedParseExpressions(p -> p);
-            return context.getNamedParseExpressions().push(new org.apache.spark.sql.catalyst.expressions.And(new GreaterThanOrEqual(value, lower), new LessThanOrEqual(value, upper)));
-        }
-
-        @Override
-        public Expression visitCidr(org.opensearch.sql.ast.expression.Cidr node, CatalystPlanContext context) {
-            analyze(node.getIpAddress(), context);
-            Expression ipAddressExpression = context.getNamedParseExpressions().pop();
-            analyze(node.getCidrBlock(), context);
-            Expression cidrBlockExpression = context.getNamedParseExpressions().pop();
-
-            ScalaUDF udf = new ScalaUDF(SerializableUdf.cidrFunction,
-                    DataTypes.BooleanType,
-                    seq(ipAddressExpression,cidrBlockExpression),
-                    seq(),
-                    Option.empty(),
-                    Option.apply("cidr"),
-                    false,
-                    true);
-
-            return context.getNamedParseExpressions().push(udf);
         }
     }
 }
